gl-renderer: Set pitch correctly for subsampled textures
Multi-plane sub-sampled textures have partial width/height, e.g.
YUV420/I420 has a full-size Y plane, followed by a half-width/height U
plane, and a half-width/height V plane.

zwp_linux_dmabuf_v1 allows clients to pass an explicit pitch for each
plane, but for wl_shm this must be inferred. gl-renderer was correctly
accounting for the width and height when subsampling, but the pitch was
being taken as the pitch for the first plane.

This does not match the requirements for GStreamer's waylandsink, in
particular, as well as other clients. Fix the SHM upload path to
correctly set the pitch for each plane, according to subsampling.

Tested with:
  $ gst-launch-1.0 videotestsrc ! waylandsink

Signed-off-by: Daniel Stone <daniels@collabora.com>
Fixes: fdeefe42418 ("gl-renderer: add support of WL_SHM_FORMAT_YUV420")
Reported-by: Fabien Lahoudere <fabien.lahoudere@collabora.co.uk>
Bugzilla: https://bugs.freedesktop.org/show_bug.cgi?id=103063
Cc: Vincent Abriou <vincent.abriou@st.com>