launcher: logind : Fix to avoid compositor to go asleep when drm device removed
drm_device_is_kms is called multiple times for each DRM device node. When
weston_launcher_close is called on a non-kms DRM device, it will deactivate
the session and cause the compositor to sleep.

Add a check for "gone" event when the device gets removed, to avoid deactivating
the compositor.

The case was reproduced on an embedded device having muliple DRM nodes on
which the order of udev_enumerate_get_list_entry was not predictable.

Upstream Status: Proposed

Signed-off-by: Prabhu Sundararaj <prabhu.sundararaj@nxp.com>
(cherry picked from commit 0a1912e121691c64b6919532180b32377ed9235a)