compositor-drm: Repaint the previous damage rectangle when current damage is zero
This issue only exists with g2d-renderer.
It need repaint with the previous damage rectangle if
current damage is zero. When app exits it has a fade
out effect with a diminishing alpha value in weston view,
but g2d-dpu doesn't support global alpha. So when the
fade-out finish without effect, it will have a zero damage,
then it need repaint with the previous damage.

Upstream-Status: Inappropriate [i.MX-specific]

Signed-off-by: Meng Mingming <mingming.meng@nxp.com>