MGS-4050 [#ccc] libweston/compositor-fbdev: Use fbGetDisplayByIndex when selecting the framebuffer
When user wants to use a different framebuffer than default he can
export a custom VSI environment variable. This is error prone and users
are not aware of it.  Instead users can use --device arg to pass the
correct framebuffer directly. Problem is we didn't use
fbGetDisplayByIndex to select the proper display/framebuffer. Adress
this issues by using fbGetDisplayByIndex.

Signed-off-by: Marius Vlad <marius-cristian.vlad@nxp.com>
Reported-by: Leonard Crestez <leonard.crestez@nxp.com>