compositor-drm: Met process stuck when do GPU PM Test
drm master was set by systemd in PM test, try to set the master back
This is a workaround patch, we should detect the EPERM error when we
try to commit,
cancel the repaint, and wait for the session to be deactivated. When
the session is reactivated,
we should make sure that repaint still works from there.

Upstream-Status: Inappropriate [i.MX-specific]

Signed-off-by: Wujian sun <wujian.sun@nxp.com>
(cherry picked from commit b76660e48717d0e34c1a8f95a91ecb3c7cccba32)