backend-drm: Improve the way to find plane for view
1. remove the restriction on underlay planes when finding plane.
Because the view on the underlay plane can be displayed by drawing
a through hole on primary plane, so we can try underlay planes.

2. Add step to check if the view is assigned on underlay plane, When
it is successfully placed on a HW plane. Because we need to set the
underlay view pnode->need_through_hole to true so that gl-renderer
will draw a hole for it  when repainting.

Upstream-Status: Pending [i.MX-specific]
Signed-off-by: Chao Guo <chao.guo@nxp.com>
(cherry picked from commit 4249bd3631f61b93baa5622b63be80254d64d419)