kms: fix hotplug HDMI meet black screen when play video on multi-screen
When a video is displayed on multiple screens and one of the HDMI connectors
is unplugged, Weston has a certain probability of getting stuck.

The reason is that there is a certain probability that the unplugged output
will not be destroyed before atomic commit, so the corresponding CRTC will
not be disabled, resulting in atomic commit failure and Weston being unable
to schedule the next repaint time.

A judgment is added in drm_pending_state_apply_atomic() function that before
atomic commit, the CRTC corresponding to outputs that have already been
unplugged but not yet destroyed will be disabled.

Upstream-Status: Pending

Signed-off-by: Chao Guo <chao.guo@nxp.com>