libweston: Add a new member to the weston_paint_node structure
Add a boolean member named need_through_hole in weston_paint_node,
which is used to indicate whether the renderer should draw through
hole on primary plane when rendering.

For paint node whose view are placed on the underlay plane, this
member should be set to true, otherwise it is false.

Upstream-Status: Pending [i.MX-specific]
Signed-off-by: Chao Guo <chao.guo@nxp.com>