backend-drm: make the renderer produce a non-opaque image
1. When getting fb, according to whether there is holes on primary plane,
let the renderer generate a non-opaque or opaque image, so that the view
under the hole can be displayed.

2. When reusing drm_fb, check whether the format of drm_fb meets the
transparency requirements, if not, modify the format and re-add fb.
Because the format of the image will change depending on whether
there are holes on primary plane.

Upstream-Status: Pending [i.MX-specific]
Signed-off-by: Chao Guo <chao.guo@nxp.com>
(cherry picked from commit c32d39fc96c66db6d31a09ebd46756443486d577)