gl-renderer: Draw holes on primary plane for the view on underlay planes
1. Added a new function for drawing through holes on primary plane. In
this function, a gl-shader is configured to draw an opaque solid color
area with a color value of (0, 0, 0, 0). The location of this area is
determined by the target view.

2. Add this new function to repaint_views(). When gl-renderer repaints
views, it will draw through holes for the needed views according to
need_through_hole, so that the underlay views can display.

Upstream-Status: Pending [i.MX-specific]
Signed-off-by: Chao Guo <chao.guo@nxp.com>