LF-10886 g2d-renderer: Add rotation function for video surface
When Wayland client (such as waylandsink) sets the rotation angle for video
playback, the video plays abnormally on g2d platform - the viewport of the
video is rotated, but the video content is not.

This is because the rotation of the dstSurface and srcSurface are different
when drawing the video view.

      dstsurface_roation  = view_rotation + output_rotation
      srcsurface_roataion = output_roataion

Where the view_rotation is set by the Wayland client, the output_roataion is
the output rotation angle, which can be set by weston.ini.

In order to make video playback normal, the modifications are as follows:
1. Calculate correct rotation for srcSurface before g2d-renderer blit surface.
2. Optimize the srcRect clip function.
   * Determine the cropped part of srcRect based on the rotation of srcSurface.
   * When the rotation angle is 270 and 180, add the judgment of whether dstRect
     exceeds the right and lower boundaries.

Upstream-Status: Inappropriate [i.MX-specific]

Signed-off-by: Chao Guo <chao.guo@nxp.com>
(cherry picked from commit ddacf3d46740b187c7a2d4ed5773853ee8796663)