LF-10887 g2d-renderer: Restore the correct src width and height based on rotation
In waylansink, if user sets the video surface rotation to 90/270, the values
of src_width and src_height will be exchanged and submitted to the weston
compositor.

               src_width = video_height
               src_height= video_width

g2d-renderer will set src_width and src_height to the right and bottom boundaries
of source buffer, when blitting buffer to the destination region.

Since the video buffer is still stored as video_width X video_height, g2d-renderer
will read the wrong buffer into the destination region.

Therefore, we need to restore src_width and src_height according to the rotation
method before setting the boundary of source buffer.

Upstream-Status: Inappropriate [i.MX-specific]

Signed-off-by: Chao Guo <chao.guo@nxp.com>