LF-13959 compositor: add frame callback for occluded pnode
To avoid dropping buffer, Waylandsink needs Weston to send frame callback
to inform it that the current frame painting is complete and start drawing
the next frame.

When the video is occluded, in order for the waylandsink to receive the
frame callback, the pnode callback need to be added to the current output
frame callbacks list.

Upstream-Status: Inappropriate [i.MX-specific]
Signed-off-by: Chao Guo <chao.guo@nxp.com>